// next-app/src/lib/types.ts
export interface DungCuModel {
  id_DungCu: number;
  tenDungCu: string;
  gia: number;
  giaKhuyenMai: number | null;
  moTa: string | null;
  hinhAnh: string | null;
  soLuong: number;
  daXoa: number;
  thongTin: string | null;
  giaNhap: number | null;
  id_LoaiDungCu: number | null;
  id_NhaCungCap: number | null;
  id_NhaSanXuat: number | null;
  loaiDungCu: LoaiDungCuModel | null;
  nhaCungCap: NhaCungCapModel | null;
  nhaSanXuat: NhaSanXuatModel | null;
}

export interface LoaiDungCuModel {
  id_LoaiDungCu: number;
  tenLoaiDungCu: string;
}

export interface NhaCungCapModel {
  id_NhaCungCap: number;
  tenNhaCungCap: string;
  diaChi: string | null;
  soDienThoai: string | null;
  email: string | null;
}

export interface NhaSanXuatModel {
  id_NhaSanXuat: number;
  tenNSX: string;
  diaChi: string | null;
  email: string | null;
}

export interface BannerModel {
  id_Banner: number;
  tenBanner: string | null;
  duongDan: string | null;
  lienKet: string | null;
}

export interface KhachHangModel {
  id_KhachHang: string;
  tenKhachHang: string | null;
  email: string | null;
  soDienThoai: string | null;
  diaChi: string | null;
  diemThuong: number | null;
}

export interface YeuThichModel {
  id_YeuThich: number;
  id_KhachHang: string;
  id_DungCu: number;
  sanPham: DungCuModel;
}